+++ src/components/allianceItemData.ts
