+++ src/components/subscriptions/enthusiast-subscription.vue
@@ -95,7 +95,7 @@
 </template>
 
 <script>
-    import {enthusiastItemData} from "../EnthusiastItemData";
+    import {enthusiastItemData} from "../enthusiastItemData";
 
     const decodeHtml = require('he').decode;
     function last(arr) {
